backend/requirements.txt
fastapi
uvicorn
starlette
python-dotenv
fastmcp
mcp[cli]
asyncpg
psycopg[binary]
psycopg-pool
azure-storage-blob
pymupdf
github-copilot-sdk
pydantic